pyproject.toml
[project]
name = "domino-py-iisas"
dynamic = ["version", "readme"]
description = "Fork of the original Python package for Domino."
authors = [
    { name = "Stefan Dlugolinsky", email = "stefan.dlugolinsky@savba.sk"},
    { name = "Luiz Tauffer", email = "luiz@taufferconsulting.com" },
    { name = "Vinicius Vaz", email = "vinicius@taufferconsulting.com" }
]
requires-python = ">=3.11"
keywords = ["domino", "airflow", "gui"]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: User Interfaces",

]
dependencies = [
    "pydantic==2.11.7",
    "tomli==2.0.1",
    "tomli-w==1.0.0",
    "PyYAML==6.0.2",
    "jsonschema==4.25.1",
    "click==8.2.1",
    "rich==13.9.4",
    "colorama==0.4.6",
]

[project.urls]
homepage = "https://github.com/IISAS/domino"
documentation = "https://tauffer-consulting.github.io/domino-docs/"
repository = "https://github.com/IISAS/domino"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
domino = [
    'cli/utils/config-domino-local.toml',
    'cli/utils/docker-compose.yaml',
    'custom_operators/sidecar/sidecar_lifecycle.sh',
    'custom_operators/sidecar/rclone.conf',
    'custom_operators/sidecar/fuse.conf',
    'VERSION',
]

[tool.setuptools.dynamic]
version = { attr = "domino.__version__" }
readme = { file = 'README.md', content-type = 'text/markdown' }

[project.scripts]
domino = "domino.cli.cli:cli"

[project.optional-dependencies]
cli = [
    "urllib3==2.5.0",
    "cryptography==45.0.7",
    "pyOpenSSL==25.1.0",
    "PyGithub==1.59.1",
    "docker==7.1.0",
    "kubernetes==32.0.1",
    "bottle==0.13.4",
    "requests==2.32.5",
    "python-gitlab==8.1.0"
]
airflow = [
    "apache-airflow==3.0.6",
    "apache-airflow-providers-cncf-kubernetes==10.7.0",
    "apache-airflow-providers-docker==4.4.2"
]
full = [
    "bottle==0.13.4",
    "urllib3==2.5.0",
    "cryptography==45.0.7",
    "pyOpenSSL==25.1.0",
    "PyGithub==1.59.1",
    "docker==7.1.0",
    "kubernetes==32.0.1",
    "apache-airflow==3.0.6",
    "apache-airflow-providers-cncf-kubernetes==10.7.0",
    "apache-airflow-providers-docker==4.4.2",
    "python-gitlab==8.1.0"
]
